Incorporated AVRCP into Go module
