--- devices.json.tmpl
@@ -1,3 +1,0 @@
-[
-    {"name": "NAME", "mac_address": "MAC ADDRESS"}
-]
