--- requirements.txt
@@ -1,2 +1,0 @@
-dbus-python==1.2.16
-PyGObject==3.40.1
